chore: Update KiCAD gerber and drill files
The code changes include updating the KiCAD gerber and drill files in the M01 and X01 directories. These updates ensure that the latest versions of the files are included in the repository.

Note: The recent user commits and recent repository commits do not provide any relevant information for generating the commit message.
Signed-off-by: Jose Vega <josevega234@gmail.com>